+++ graphite/systems/InputSystem.h
@@ -1,12 +1,15 @@
 #pragma once
 
 #include "core/ISystem.h"
+#include "input/InputManager.h"
 #include <Windows.h>
 
 class InputSystem : public ISystem
 {
 public:
-    void Init() override { OutputDebugStringA("[InputSystem] Initialized.\n"); };
-    void Update() override { /* OutputDebugStringA("[InputSystem] Updated.\n") */ ; };
-    void Shutdown() override { OutputDebugStringA("[InputSystem] Shutdown.\n"); };
+    void Init() override {};
+    void Update(float /*deltaTime*/) override {
+        // InputManager::Get().NewFrame();
+    };
+    void Shutdown() override {};
 };